\name{ada.weights}
\alias{ada.weights}
\title{Internal Logic Forest Functions
}
\description{This function is called by \code{\link{LBoost}}.  It is not intended to be used independently of this function.  
}
\details{Determines the weight given to each tree in an LBoost model.  These tree specific weights are used when making predictions from an LBoost model.
}
\author{Bethany Wolf wolfb@musc.edu
}
\note{This function is supplementary and not intended to be used independently of the other functions.
}
\keyword{internal}

