\name{CV.err}
\alias{CV.err}

\title{Internal Logic Forest Functions
}
\description{This function is called by \code{\link{LBoost}}.  It is not intended to be used independently of this function.  
}
\details{Estimated the error rate of an LBoost model using the test data sets generated by \code{\link{CV.data}}.
}
\author{Bethany Wolf wolfb@musc.edu
}
\note{This function is supplementary and not intended to be used independently of the other functions.
}
\keyword{internal}
