\name{CV.ids}
\alias{CV.ids}

\title{Internal Logic Forest Functions
}
\description{This function is called by \code{\link{LBoost}} and \code{\link{PlusMinus.PIimp}}.  It is not intended to be used independently of these function.  
}
\details{Determines the row number for observations included in each training and test dataset generated by \code{\link{CV.data}}.
}
\author{Bethany Wolf wolfb@musc.edu
}
\note{This function is supplementary and not intended to be used independently of the other functions.
}
\keyword{internal}
