\name{LBoost.fit}
\alias{LBoost.fit}
\docType{data}
\title{Example of a LBoost Model
}
\description{The LBoost model for data with a binary response with error following the underlying logical relationship 


L=(X4 and X5) or (X9 and X10)


The LBoost model used 5-fold cross validation and includes 100 logic regression trees.  Predictor importance and two measures of interaction (prime implicant) importance were also run as part of the model.
}
\usage{data(LBoost.fit)}
\format{An object of class \code{LBoost}.
}


\examples{
data(LBoost.fit)
}
\keyword{datasets}

