\name{LBoost.PIs}
\alias{LBoost.PIs}
\title{Internal Logic Forest Functions
}

\description{This function is called by \code{\link{LBoost}}.  It is not intended to be used independently of this function.  
}

\details{Collects a matrix indicating which prime implicants are identified by a tree in the LBoost model.  Also generates a matrix that indicates which interaction terms match each observation in the training data.
}


\author{Bethany Wolf wolfb@musc.edu
}

\note{This function is supplementary and not intended to be used independently of the other functions.
}
\keyword{internal}

