\name{LFfit1}
\alias{LFfit1}
\title{Example of a Logic Forest Model
}
\description{The Logic Forest model for data with noisy predictors following the underlying logical relation ship 


L=(X4 or X11) and X5


The logic forest includes 100 logic regression trees.
}
\usage{data(LFfit1)}
\format{An object of class \code{logforest}.
}

\examples{
data(LFfit1)
}
\keyword{datasets}


