\name{list.PIs}
\alias{list.PIs}
\title{Internal Logic Forest Functions
}
\description{Internal function called by \code{\link{subs}} and \code{\link{persist.match}}.  Not intended to be used independent of these functions.
}
\usage{
list.PIs(match.list, allPIs, preds)
}
\arguments{
  \item{match.list}{a list of interactions for which subset matches are to be identified.
}
  \item{allPIs}{a list of all interactions identified in a logic forest model.
}
  \item{preds}{number of predictors in the original dataset.
}
}

\author{Bethany Wolf wolfb@musc.edu
}

\keyword{ internal }% __ONLY ONE__ keyword per line


