\name{logforest.fit}
\alias{logforest.fit}
\docType{data}
\title{Example of a Logic Forest Model
}
\description{The Logic Forest model for data with a binary response with error following the underlying logical relationship 


L=(X4 and X5) or (X9 and X10)


The logic forest includes 100 logic regression trees.
}
\usage{data(logforest.fit)}
\format{An object of class \code{logforest}.
}

\examples{
data(logforest.fit)
}
\keyword{datasets}

