\name{Perm.PIimp}
\alias{Perm.PIimp}

\title{Internal Logic Forest Functions
}
\description{This function is called by \code{\link{LBoost}}.  It is not intended to be used independently of this function.  
}
\details{Determines the permutation based importance measure for all interactions corresponding to a specific logic regression tree within an LBoost model.
}
\author{Bethany Wolf wolfb@musc.edu
}
\note{This function is supplementary and not intended to be used independently of the other functions.
}
\keyword{internal}
