\name{Perms}
\alias{Perms}
\title{Internal Logic Forest Functions
}
\description{This function is called by \code{TTab}.  It is not intended to be used independently of this function.  
}


\references{
}
\author{Bethany Wolf wolfb@musc.edu
}

\note{This function is supplementary and not intended to be used  independently of the other functions.
}
\keyword{internal}
