\name{persist.match}
\alias{persist.match}
\title{Internal Logic Forest Functions
}
\description{Internal function called by \code{\link{persistence.prep}}.  Not intended to be used independently of this function.
}

\details{ Generates a list of all subset matches for variables or variable interactions in a logic forest model or an LBoost model.  
}


\author{Bethany Wolf wolfb@musc.edu
}


\keyword{internal}

