\name{persistence.prep}
\alias{persistence.prep}
\title{Internal Logic forest Functions
}
\description{Internal function called by \code{\link{persistence.plot}}.  Not intended to be used independt of this function.
}
\usage{
persistence.prep(fit, preds, PI)
}
\arguments{
  \item{fit}{ an object of class \code{logforest}.
}
  \item{preds}{ a numeric value representing the number of predictors in the original data used to construct the logic forest model.
}
  \item{PI}{ a charachter string for the variable or interaction term to be considered for matching
}
}


\author{Bethany Wolf wolfb@musc.edu
}

\keyword{internal}


