\name{pimp.import}
\alias{pimp.import}
\title{Internal Logic Forest Functions
}
\description{This function is called by \code{\link{logforest}}.  It is not intended to be used independently of this function.  
}
\details{Determines the permutation based importance measure for all variables and variable interactions corresponding to a specific logic regression tree within a logic forest.
}
\author{Bethany Wolf wolfb@musc.edu
}
\note{This function is supplementary and not intended to be used independently of the other functions.
}
\keyword{internal}

