\name{prime.imp}
\alias{prime.imp}
\title{Internal Logic Forest Functions
}
\description{This function is called by \code{pimp.import} and \code{\link{LBoost.PIs}}.  It is not intended to be used independently of these function.  
}
\details{ It generates a list of all variables and variable interactions identified by a specific logic regression tree with in a logic forest model or and LBoost model.
}

\author{Bethany Wolf   wolfb@musc.edu
}
\note{This function is supplementary and not intended to be used  independently of the other functions.
}
\keyword{internal}


