\name{proportion.positive}
\alias{proportion.positive}
\title{Internal Logic Forest Functions
}
\description{This function is called by \code{predict.logforest}.  It is not intended to be used independetly of this function.  It determined the proportion of logic regression trees within a logic forest that predict a class of one for new observations.  Additionally it returns predicted values for new observations.
}

\references{
}
\author{Bethany Wolf wolfb@musc.edu
}
\note{This function is supplementary and not intended to be used independently of the other functions.
}
\keyword{internal}