\name{TTab}
\alias{TTab}
\title{Internal Logic Forest Functions
}
\description{This function is called by \code{prime.imp}.  It is not intended to be used independently of this function.  
}

\details{Generates a matrix of binary predictor values corresponding to a prediction of one for a specific logic regression tree within a logic forest.  Only those predictors included in the logic regression tree are included in the matrix. 
}

\references{
}
\author{Bethany Wolf wolfb@musc.edu
}
\note{This function is internal and not intended to be used  independently of the other functions.
}

\keyword{internal}

