\name{LF.data}
\alias{LF.data}
\docType{data}
\title{Example Data for Logic Forest
}

\usage{LF.data}

\description{LF.data has 200 observations and 21 predictors.  Columns 1,2,....,50 are binary predictors.  Column 51 is the binary response.  Each predictor is simulated as an intependent Bernoulli random variable with success probability 0.5.  The response variable is simulated from the model 

Y=(X4 or X11) and X5
}

\examples{
data(LF.data)
}
\keyword{datasets}
