\name{LogicForest-package}
\alias{LogicForest-package}
\alias{LogicForest}
\docType{package}
\title{
Logic Forest
}
\description{
Two classification ensemble methods based on logic regression models.  Logforest uses a bagging approach to contruct an ensemble of logic regression models.  LBoost uses a combination of boosting and cross-validation to construct and ensemble of logic regression models.  Both methods are used for classification of binary responses based on binary predictors and for identification of important variables and variable interactions predictive of a binary outcome.

}
\details{
\tabular{ll}{
Package: \tab LogicForest\cr
Type: \tab Package\cr
Version: \tab 2.0.0\cr
Date: \tab 2011-09-20\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}


}
\author{
Bethany Wolf

Maintainer: Bethany Wolf <wolfb@musc.edu>
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{

}

