\name{LogicForest-package}
\alias{LogicForest-package}
\alias{LogicForest}
\docType{package}
\title{
Logic Forest
}
\description{
Classification ensemble based on logic regression trees.  Used for classification of binary responses based on binary predictors and for identification of important variables and variable interactions predictive of a binary outcome.

}
\details{
\tabular{ll}{
Package: \tab LogicForest\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-07-07\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}


}
\author{
Bethany Wolf

Maintainer: Bethany Wolf <wolfb@musc.edu>
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{

}
