\name{pimp.mat}
\alias{pimp.mat}
\title{Internal Logic Forest Functions
}
\description{This function is called by pimp.import  It is not intended to be used independently of this function.  
}

\references{
}
\author{Bethany Wolf wolfb@musc.edu
}
\note{This function is internal and not intended to be used independently of the other functions.
}
\keyword{internal}
