\name{prime.imp}
\alias{prime.imp}
\title{Internal Logic Forest Functions
}
\description{This function is called by \code{pimp.import}.  It is not intended to be used independently of this function.  It generates a list of all predictors and prime implicants identified by a specific logic regression tree with in a logic forest.
}

\author{Bethany Wolf wolfb@musc.edu
}
\note{This function is supplementary and not intended to be used  independently of the other functions.
}
\keyword{internal}