\name{subs}
\alias{subs}
\title{Internal Logic Forest Functions
}
\description{This function is called by \code{persistence.prep}.  It is not intended to be used independently of this function.  
}
\author{Bethany Wolf wolfb@musc.edu
}
\note{This function is internal and not intended to be used  independently of the other functions.
}
\keyword{internal}

